pyproject.toml
[tool.poetry]
name = "autoqml-lib"
version = "0.1.0"
description = ""
authors = ["marc <marc.zoeller@usu.com>", "Milan Marinov <milan.marinov@usu.com"]
readme = "README.md"
packages = [{include = "autoqml_lib"}]

[tool.poetry.dependencies]
python = "^3.9, <3.11"
pandas = "^2.0.2"
ray = {extras = ["tune"], version = "^2.4.0"}
squlearn = ">=0.8.0"
rustworkx = ">=0.12.0"
tune-sklearn = "^0.4.5"
scikit-learn = "^1.2.2"
protobuf = "3.20"
umap-learn = "^0.5.3"
optuna = "^3.3.0"
matplotlib = "^3.8.1"
dill = "^0.3.7"
lightning = "^2.2.0.post0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[tool.pytest.ini_options]
filterwarnings = [
    # "ignore",  # would ignore all warnings
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.poetry.group.dev.dependencies]
debugpy = "^1.6.7"
yapf = "^0.40.1"
mypy = "^1.4.1"
pylint = "^2.17.4"
python-lsp-server = {extras = ["pylint", "yapf"], version = "^1.7.4"}
pylsp-mypy = "^0.6.7"
pyls-isort = "^0.2.2"
flake8 = "^6.0.0"
ipython = "^8.14.0"
jupytext = "^1.15.2"
jupyterlab = "^4.0.8"
ipywidgets = "^8.1.1"
setuptools = "<70.0"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
warn_unused_ignores = true
show_error_context = true
show_column_numbers = true
show_error_codes = true
pretty = true

[tool.pylsp-mypy]
enabled = true
live_mode = true
strict = false

[tool.yapf]
based_on_style = "facebook"
column_limit = 79

[tool.pylint.'FORMAT']
indent-after-paren=4
indent-string='    '
max-line-length=79
docstring-min-length=5

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
